pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trading-engine"
version = "0.2.0"
description = "Production-grade ML trading engine with XGBoost signal generation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "pyarrow>=10.0.0",
    "xgboost>=1.7.0",
    "optuna>=3.0.0",
    "scikit-learn>=1.2.0",
    "joblib>=1.2.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "pydantic>=2.0.0",
    "sse-starlette>=1.6.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.18",
]

[project.optional-dependencies]
ibkr = ["ib_insync>=0.9.86"]
dev  = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "ruff>=0.1.0"]

[project.scripts]
trading-engine = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
